red: push-tap marks the deep-linked application read (tc-0sfx.3)
